Folder rename from 'components' to 'pages'
