+++ src/pages/currency-chart/index.js
