+++ src/pages/foreign-currencies-rates/index.js
