+++ src/pages/foreign-currencies-rates/style.css
