+++ src/pages/gold-prices/index.js
